backend/pyproject.toml
[project]
name = "dbt-ui-backend"
version = "0.1.0"
description = "FastAPI backend for dbt-ui, an open-source local UI wrapper around dbt-core"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "aiosqlite>=0.20.0",
    "sse-starlette>=2.1.3",
    "watchfiles>=0.24.0",
    "dbt-artifacts-parser>=0.8.1",
    "pyyaml>=6.0.2",
    "ptyprocess>=0.7.0",
    "structlog>=24.4.0",
    "python-multipart>=0.0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.2",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
